データの取得を service に分離
