+++ src/app/todo/todo-list/todo-list.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-
+import { DataService } from 'src/app/data.service';
+import { TodoCategory } from 'src/app/data.service';
 @Component({
   selector: 'app-todo-list',
   templateUrl: './todo-list.component.html',
@@ -8,47 +8,14 @@ import { HttpClient } from '@angular/common/http';
 })
 
 export class TodoListComponent {
-  constructor(private http: HttpClient) {}
+  constructor(private dataService: DataService) {}
 
   todoCategoryList: TodoCategory[] = [];
 
-  API = 'http://localhost:9000/todo/list'
-
   ngOnInit(): void {
-    this.getTodoCategoryList();
-  }
-
-  getTodoCategoryList(): void {
-    this.http.get<TodoCategory[]>(this.API).subscribe(
-      (response) => {
-        this.todoCategoryList = response;
-        return response;
-      },
-      (error) => {
-        console.error(error);
-      }
-    )
+    this.dataService.getTodoCategoryList().subscribe(data => {
+      this.todoCategoryList = data;
+    });
   }
 }
 
-interface State {
-  code: Number,
-  name: String
-}
-interface Todo {
-  id:         Number,
-  categoryId: Number,
-  title:      String,
-  body:       String,
-  state:      State
-}
-interface Category {
-  id:    Number,
-  name:  String,
-  slug:  String,
-  color: String
-}
-interface TodoCategory {
-  todo: Todo,
-  category: Category
-}
